Collect all data migration code in own folder
