+++ src/main/apps/sources.test.ts
@@ -4,14 +4,7 @@
  * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
  */
 
-import { existsSync } from 'fs';
-
-import { readSchemedJsonFile, writeSchemedJsonFile } from '../fileUtil';
-import {
-    migrateSourcesJson,
-    newWithdrawnJson,
-    oldSourcesJsonPath,
-} from './sources';
+import { newWithdrawnJson } from './sources';
 
 jest.mock('fs');
 jest.mock('../fileUtil');
@@ -44,59 +37,4 @@ test('computing the new list of withdrawn apps', () => {
         'http://example.org/oldWithdrawnApp.json',
         'http://example.org/newlyWithdrawnApp.json',
     ]);
-});
-
-describe('migrating sources.json', () => {
-    const mockedExistsSync = jest.mocked(existsSync);
-    const mockedReadSchemedJsonFile = jest.mocked(readSchemedJsonFile);
-    const mockedWriteSchemedJsonFile = jest.mocked(writeSchemedJsonFile);
-
-    beforeEach(() => {
-        jest.resetAllMocks();
-    });
-
-    it('should migrate sources.json to sources-versioned.json', () => {
-        const oldSourceJsonContent = {
-            source1: 'http://source1.com/apps.json',
-            source2: 'http://source2.com/source.json',
-        };
-        const newSourceVersionedJsonContent = {
-            v1: [
-                { name: 'source1', url: 'http://source1.com/source.json' },
-                { name: 'source2', url: 'http://source2.com/source.json' },
-            ],
-        };
-
-        // Arrange
-        mockedExistsSync.mockImplementation(
-            path => path === oldSourcesJsonPath()
-        );
-        mockedReadSchemedJsonFile.mockReturnValue(oldSourceJsonContent);
-
-        // Act
-        migrateSourcesJson();
-
-        // Assert
-        expect(mockedWriteSchemedJsonFile).toBeCalledWith(
-            expect.anything(),
-            expect.anything(),
-            newSourceVersionedJsonContent
-        );
-    });
-
-    it('should do nothing if sources.json does not exist', () => {
-        mockedExistsSync.mockReturnValue(false);
-
-        migrateSourcesJson();
-
-        expect(mockedReadSchemedJsonFile).not.toBeCalled();
-    });
-
-    it('should do nothing if sources-versioned.json already exists', () => {
-        mockedExistsSync.mockReturnValue(true);
-
-        migrateSourcesJson();
-
-        expect(mockedReadSchemedJsonFile).not.toBeCalled();
-    });
 });
