+++ src/main/index.ts
@@ -10,7 +10,7 @@ import './setUserDataDir';
 import { initialize as initializeElectronRemote } from '@electron/remote/main';
 import telemetry from '@nordicsemiconductor/pc-nrfconnect-shared/src/telemetry/telemetry';
 
-import { migrateSourcesJson } from './apps/sources';
+import { migrateSourcesJson } from './apps/dataMigration/migrateSourcesJson';
 import configureElectronApp from './configureElectronApp';
 import initNrfUtilProxyEnv from './initNrfUtilProxyEnv';
 import registerIpcHandler from './registerIpcHandler';
